Name dart_runner processes after the app they're running
Currently in `ps` all apps running under the Dart content handler show
up as `dart_runner`. This makes them hard to identify. This change names
the process after the final path component URL of the application that's
launched, prefixed with "dart:".

In theory we could run more than one Dart application in a single
content handler process, but we don't right now. If we do this logic
would have to change.

Change-Id: I28ef03fc50cebbff0fcacda6b24400be2956f50f